+++ lib/crowi/express-init.js
@@ -11,14 +11,13 @@ module.exports = function(crowi, app) {
   var basicAuth = require('basic-auth-connect')
   var flash = require('connect-flash')
   var cons = require('consolidate')
-  var swig = require('swig')
   var i18next = require('i18next')
   var i18nFsBackend = require('i18next-node-fs-backend')
   var i18nSprintf = require('i18next-sprintf-postprocessor')
   var i18nMiddleware = require('i18next-express-middleware')
   var i18nUserSettingDetector = require('../util/i18nUserSettingDetector')
   var env = crowi.node_env
-  var middleware = require('../util/middlewares')
+  const middlewares = require('../middlewares')(crowi, app)
 
   var User = crowi.model('User')
 
@@ -106,10 +105,10 @@ module.exports = function(crowi, app) {
 
   app.use(flash())
 
-  app.use(middleware.swigFilters(app, swig))
-  app.use(middleware.swigFunctions(crowi, app))
+  app.use(middlewares.SwigFilters)
+  app.use(middlewares.SwigFunctions)
 
-  app.use(middleware.loginChecker(crowi, app))
+  app.use(middlewares.LoginChecker)
 
   app.use(i18nMiddleware.handle(i18next))
   app.use((req, res, next) => {
